refactoring: moved  to  class
