+++ tests/project_test.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 import os
 import unittest
-from blameandshame.base import  Project
+from blameandshame.base import Project, Change
 
 
 class ProjectTestCase(unittest.TestCase):
@@ -18,6 +18,33 @@ class ProjectTestCase(unittest.TestCase):
                          os.path.join(repos_dir, 'dep'))
         self.assertEqual(Project._url_to_path('https://github.com/opencv/opencv.git'),
                          os.path.join(repos_dir, 'opencv'))
+
+
+    def test_files_in_commit(self):
+        project = Project.from_url('https://github.com/google/protobuf')
+        self.assertEqual(project.files_in_commit('baed06e'),
+                         frozenset(['objectivec/GPBCodedOutputStream.m']))
+        self.assertEqual(project.files_in_commit('949596e'),
+                         frozenset(['objectivec/GPBMessage.m']))
+        self.assertEqual(project.files_in_commit('cd5f49d', {Change.MODIFIED}),
+                         frozenset(['ruby/travis-test.sh',
+                                    'ruby/ext/google/protobuf_c/protobuf.c',
+                                    'ruby/ext/google/protobuf_c/defs.c',
+                                    'ruby/Rakefile',
+                                    'Makefile.am',
+                                    '.gitignore',
+                                    'ruby/ext/google/protobuf_c/protobuf.h']))
+        self.assertEqual(project.files_in_commit('cd5f49d', {Change.ADDED}),
+                         frozenset(['ruby/tests/gc_test.rb']))
+        self.assertEqual(project.files_in_commit('cd5f49d'),
+                         frozenset(['ruby/travis-test.sh',
+                                    'ruby/ext/google/protobuf_c/protobuf.c',
+                                    'ruby/ext/google/protobuf_c/defs.c',
+                                    'ruby/Rakefile',
+                                    'Makefile.am',
+                                    '.gitignore',
+                                    'ruby/ext/google/protobuf_c/protobuf.h',
+                                    'ruby/tests/gc_test.rb']))
 
 
 if __name__ == '__main__':
