+++ tests/util_test.py
@@ -1,8 +1,7 @@
 #!/usr/bin/env python3
 import unittest
 from blameandshame.base import  Project, Change
-from blameandshame.util import  files_in_commit, \
-                                lines_modified_by_commit, \
+from blameandshame.util import  lines_modified_by_commit, \
                                 authors_of_file, \
                                 commits_to_file, \
                                 commits_to_line, \
@@ -54,36 +53,6 @@ class UtilTestCase(unittest.TestCase):
                   ['Paul Yang', 'michaelbausor', 'Brent Shaffer'])
         check_one(repo, 'php/composer.json', '6b27c1f',
                   ['Paul Yang'])
-
-
-    def test_files_in_commit(self):
-        project = Project.from_url('https://github.com/google/protobuf')
-        repo = project.repo
-        self.assertEqual(files_in_commit(repo, 'baed06e'),
-                         frozenset(['objectivec/GPBCodedOutputStream.m']))
-        self.assertEqual(files_in_commit(repo, '949596e'),
-                         frozenset(['objectivec/GPBMessage.m']))
-        self.assertEqual(files_in_commit(repo, 'cd5f49d', {Change.MODIFIED}),
-                         frozenset(['ruby/travis-test.sh',
-                                    'ruby/ext/google/protobuf_c/protobuf.c',
-                                    'ruby/ext/google/protobuf_c/defs.c',
-                                    'ruby/Rakefile',
-                                    'Makefile.am',
-                                    '.gitignore',
-                                    'ruby/ext/google/protobuf_c/protobuf.h']))
-        self.assertEqual(files_in_commit(repo, 'cd5f49d', {Change.ADDED}),
-                         frozenset(['ruby/tests/gc_test.rb']))
-        self.assertEqual(files_in_commit(repo, 'cd5f49d'),
-                         frozenset(['ruby/travis-test.sh',
-                                    'ruby/ext/google/protobuf_c/protobuf.c',
-                                    'ruby/ext/google/protobuf_c/defs.c',
-                                    'ruby/Rakefile',
-                                    'Makefile.am',
-                                    '.gitignore',
-                                    'ruby/ext/google/protobuf_c/protobuf.h',
-                                    'ruby/tests/gc_test.rb']))
-        # note: `ruby/tests/gc_test.rb` is added and thus should not be
-        #       considered as 'modified'.
 
 
     def test_lines_modified_by_commit(self):
